stripe_project/requirements.txt
Django==3.2
djangorestframework==3.13.1
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
urllib3==1.26.12
python-dotenv==0.20.0
psycopg2-binary
gunicorn==20.1.0